+++ server/src/playlist/playlist.controller.ts
@@ -26,11 +26,15 @@ import {
 	Post,
 	Put,
 	Query,
+	Req,
 } from "@nestjs/common";
 import { ApiOperation, ApiPropertyOptional, ApiTags } from "@nestjs/swagger";
+import { User } from "@prisma/client";
 import { IsOptional } from "class-validator";
 import AlbumService from "src/album/album.service";
 import AlbumQueryParameters from "src/album/models/album.query-parameters";
+import { Role } from "src/authentication/roles/roles.decorators";
+import Roles from "src/authentication/roles/roles.enum";
 import IdentifierParam from "src/identifier/identifier.pipe";
 import TransformIdentifier from "src/identifier/identifier.transform";
 import { PaginationParameters } from "src/pagination/models/pagination-parameters";
@@ -68,20 +72,23 @@ export default class PlaylistController {
 		summary: "Get one playlist",
 	})
 	@Get(":idOrSlug")
+	@Role(Roles.Default)
 	@Response({ handler: PlaylistResponseBuilder })
 	async get(
 		@IdentifierParam(PlaylistService)
 		where: PlaylistQueryParameters.WhereInput,
 		@RelationIncludeQuery(PlaylistQueryParameters.AvailableAtomicIncludes)
 		include: PlaylistQueryParameters.RelationInclude,
+		@Req() req: Express.Request,
 	) {
-		return this.playlistService.get(where, include);
+		return this.playlistService.get(where, this._getUserId(req), include);
 	}
 
 	@ApiOperation({
 		summary: "Get many playlists",
 	})
 	@Get()
+	@Role(Roles.Default)
 	@Response({ handler: PlaylistResponseBuilder, type: ResponseType.Page })
 	async getMany(
 		@Query() selector: Selector,
@@ -90,9 +97,11 @@ export default class PlaylistController {
 		include: PlaylistQueryParameters.RelationInclude,
 		@Query()
 		paginationParameters: PaginationParameters,
+		@Req() req: Express.Request,
 	) {
 		return this.playlistService.getMany(
 			selector,
+			this._getUserId(req),
 			sort,
 			paginationParameters,
 			include,
@@ -104,6 +113,7 @@ export default class PlaylistController {
 		description: "Entries as song with an 'entryId' field",
 	})
 	@Get(":idOrSlug/entries")
+	@Role(Roles.Default)
 	@Response({
 		handler: PlaylistEntryResponseBuilder,
 		type: ResponseType.Page,
@@ -115,9 +125,11 @@ export default class PlaylistController {
 		where: PlaylistQueryParameters.WhereInput,
 		@RelationIncludeQuery(SongQueryParameters.AvailableAtomicIncludes)
 		include: SongQueryParameters.RelationInclude,
+		@Req() req: Express.Request,
 	) {
 		return this.playlistService.getEntries(
 			where,
+			this._getUserId(req),
 			paginationParameters,
 			include,
 		);
@@ -125,53 +137,71 @@ export default class PlaylistController {
 
 	@ApiOperation({
 		summary: "Create playlist",
+		description: "User has to be authenticated",
 	})
 	@Post()
+	@Role(Roles.User)
 	@Response({ handler: PlaylistResponseBuilder })
 	async createPlaylist(
 		@Body()
 		creationDto: CreatePlaylistDTO,
+		@Req() req: Express.Request,
 	) {
-		return this.playlistService.create(creationDto);
+		return this.playlistService.create({
+			...creationDto,
+			ownerId: this._getUserId(req)!,
+		});
 	}
 
 	@ApiOperation({
 		summary: "Update playlist",
 	})
 	@Put(":idOrSlug")
+	@Role(Roles.User)
 	@Response({ handler: PlaylistResponseBuilder })
 	async updatePlaylist(
 		@Body()
 		updateDto: UpdatePlaylistDTO,
 		@IdentifierParam(PlaylistService)
 		where: PlaylistQueryParameters.WhereInput,
+		@Req() req: Express.Request,
 	) {
-		return this.playlistService.update(updateDto, where);
+		return this.playlistService.update(
+			updateDto,
+			where,
+			this._getUserId(req)!,
+		);
 	}
 
 	@ApiOperation({
 		summary: "Delete playlist",
+		description: "User has to be authenticated",
 	})
 	@Delete(":idOrSlug")
+	@Role(Roles.User)
 	async delete(
 		@IdentifierParam(PlaylistService)
 		where: PlaylistQueryParameters.WhereInput,
+		@Req() req: Express.Request,
 	) {
-		await this.playlistService.delete(where);
+		await this.playlistService.delete(where, this._getUserId(req)!);
 	}
 
 	@ApiOperation({
 		summary: "Add song to playlist",
 	})
+	@Role(Roles.User)
 	@Post(":idOrSlug/entries")
 	async addSongToPlaylist(
 		@IdentifierParam(PlaylistService)
 		where: PlaylistQueryParameters.WhereInput,
 		@Body()
 		playlistEntryDTO: CreatePlaylistEntryDTO,
+		@Req() req: Express.Request,
 	) {
 		await this.playlistService.addSong(
 			{ id: playlistEntryDTO.songId },
+			this._getUserId(req)!,
 			where,
 		);
 	}
@@ -179,14 +209,20 @@ export default class PlaylistController {
 	@ApiOperation({
 		summary: "Reorder entries in playlist",
 	})
+	@Role(Roles.User)
 	@Put(":idOrSlug/entries/reorder")
 	async moveEntryInPlaylist(
 		@Body()
 		{ entryIds }: ReorderPlaylistDTO,
 		@IdentifierParam(PlaylistService)
 		where: PlaylistQueryParameters.WhereInput,
+		@Req() req: Express.Request,
 	) {
-		await this.playlistService.reorderPlaylist(where, entryIds);
+		await this.playlistService.reorderPlaylist(
+			where,
+			this._getUserId(req)!,
+			entryIds,
+		);
 	}
 
 	@ApiOperation({
@@ -194,10 +230,16 @@ export default class PlaylistController {
 		description: "This will delete a song from the playlist",
 	})
 	@Delete("entries/:id")
+	@Role(Roles.User)
 	async deleteEntryInPlaylist(
 		@Param("id", new ParseIntPipe())
 		entryId: number,
+		@Req() req: Express.Request,
 	) {
-		await this.playlistService.removeEntry(entryId);
+		await this.playlistService.removeEntry(entryId, this._getUserId(req)!);
+	}
+
+	private _getUserId(req: Express.Request): number | null {
+		return (req.user as User)?.id ?? null;
 	}
 }
